Heart Failure Analysis
Heart failure is a leading cause of morbidity and mortality in the United States. Accurately predicting the risk of death within one year following hospitalization is crucial for personalized care, resource allocation, and improving patient outcomes. This study explored the application of machine learning to develop a predictive model for 1-year mortality in heart failure patients.